requirements.txt
scapy
datetime
speedtest-cli
pytz
pymongo